Move internal package into the App struct so it can access the logger
